+++ src/widgets/history/ui/Title.tsx
@@ -1,0 +1,13 @@
+import '../styles/Title.css';
+
+export function HistoryTitle() {
+  return (
+    <div className='history__title'>
+      <div className='history__title__description'>
+        <hr />
+        <p>COMPANY HISTORY</p>
+      </div>
+      <h2>걸어온 길</h2>
+    </div>
+  );
+}
