+++ src/widgets/history/ui/book/Cover.tsx
@@ -1,0 +1,1 @@
+export function BookCover() {}
